change function to fixture
