+++ tests/integration/test_remoted/test_agent_communication/test_invalid_protocols_communication.py
@@ -5,17 +5,13 @@
 """
 
 import pytest
-import time
 
 from pathlib import Path
-from wazuh_testing.tools.simulators.agent_simulator import connect
 from wazuh_testing.utils.configuration import get_test_cases_data, load_configuration_template
 from wazuh_testing.modules.remoted.configuration import REMOTED_DEBUG
 from wazuh_testing.constants.paths.sockets import ANALYSISD_QUEUE_SOCKET_PATH
 from wazuh_testing.constants.daemons import ANALYSISD_DAEMON
-from wazuh_testing.tools.thread_executor import ThreadExecutor
 from wazuh_testing.tools.mitm import ManInTheMiddle
-from wazuh_testing.utils import callbacks
 
 from . import CONFIGS_PATH, TEST_CASES_PATH
 
@@ -33,9 +29,6 @@ daemons_handler_configuration = {'all_daemons': True}
 
 local_internal_options = {REMOTED_DEBUG: '2'}
 
-injectors = []
-search_patterns = []
-
 
 receiver_sockets_params = [(ANALYSISD_QUEUE_SOCKET_PATH, 'AF_UNIX', 'UDP')]
 mitm_analysisd = ManInTheMiddle(address=ANALYSISD_QUEUE_SOCKET_PATH, family='AF_UNIX', connection_protocol='UDP')
@@ -47,8 +40,8 @@ receiver_sockets, monitored_sockets = None, None  # Set in the fixtures
 # Test function.
 @pytest.mark.parametrize('test_configuration, test_metadata',  zip(test_configuration, test_metadata), ids=cases_ids)
 def test_invalid_protocols_communication(test_configuration, test_metadata, configure_local_internal_options, truncate_monitored_files,
-                            set_wazuh_configuration, configure_sockets_environment_module, daemons_handler, simulate_agents,
-                       connect_to_sockets_module, waiting_for_analysisd_startup):
+                            set_wazuh_configuration, configure_sockets_environment_module, daemons_handler,simulate_agents,
+                       connect_to_sockets_module, waiting_for_analysisd_startup, validate_agent_manager_protocol_communication):
 
     '''
     description: Check agent-manager communication with several agents simultaneously via TCP, UDP or both.
@@ -88,61 +81,12 @@ def test_invalid_protocols_communication(test_configuration, test_metadata, conf
         - waiting_for_analysisd_startup:
             type: fixture
             brief: Wait until the 'wazuh-analysisd' has begun and the 'alerts.json' file is created.
+        - validate_agent_manager_protocol_communication
+            type: fixture
+            brief: connect agent , launch thread and send events
     '''
 
-    agent = simulate_agents[0]
+    protocol = test_metadata['protocol']
+    manager_port = test_metadata['port']
 
-
-    def validate_agent_manager_protocol_communication(protocol='TCP', manager_port=1514):
-
-
-        def send_event(event, protocol, manager_port, agent):
-            """Send an event to the manager"""
-
-            sender, injector = connect(agent, manager_port = manager_port, protocol = protocol, wait_status= '' if protocol == 'UDP' else 'active' )
-            sender.send_event(event)
-            injectors.append(injector)
-            return injector
-
-
-        # Generate custom events for each agent
-        search_pattern = f"test message from agent {agent.id}"
-        agent_custom_message = f"1:/test.log:Feb 23 17:18:20 manager sshd[40657]: {search_pattern}"
-        event = agent.create_event(agent_custom_message)
-
-        # Create sender event threads
-        send_event_thread = ThreadExecutor(send_event, {'event': event, 'protocol': protocol,
-                                                        'manager_port': manager_port, 'agent': agent})
-
-            # If protocol is TCP, then just send the message as the attempt to establish the connection will fail.
-        if protocol == 'TCP':
-            send_event_thread.start()
-            send_event_thread.join()
-        else:  # If protocol is UDP, then monitor the  socket queue to verify that the event has not been received.
-
-            callback = callbacks.generate_callback(fr"{search_pattern}")
-            monitored_sockets[0].start(callback=callback)
-            assert monitored_sockets[0].callback_result
-
-
-            # Wait until socket monitor is fully initialized
-            time.sleep(5)
-
-            send_event_thread.start()
-            send_event_thread.join()
-
-            time.sleep(5)
-        for injector in injectors:
-            injector.stop_receive()
-
-
-    if test_metadata['protocol'] == 'TCP':
-        # Check that the connection is not established
-        with pytest.raises(ConnectionRefusedError):
-            validate_agent_manager_protocol_communication(protocol=test_metadata['protocol'], manager_port=test_metadata['port'])
-            raise ValueError('The manager has established a TCP connection when only UDP is allowed.')
-    else:
-        # Check that no event is received in the socket queue
-        with pytest.raises(AssertionError):
-            validate_agent_manager_protocol_communication(protocol=test_metadata['protocol'], manager_port=test_metadata['port'])
-            raise ValueError('The manager has received an event from a protocol not allowed.')
+    validate_agent_manager_protocol_communication(monitored_sockets, simulate_agents, protocol, manager_port)
